Move router module
