+++ assets/components/proxied-imports/router.js
