+++ assets/components/src/tabbed-navigation/index.js
@@ -11,7 +11,7 @@ import { Component } from '@wordpress/element';
  * External dependencies.
  */
 import classNames from 'classnames';
-import Router from '../router'
+import Router from '../../proxied-imports/router'
 
 /**
  * Internal dependencies.
