+++ assets/wizards/advertising/index.js
@@ -18,7 +18,7 @@ import HeaderIcon from '@material-ui/icons/FeaturedVideo';
  * Internal dependencies.
  */
 import { Card, Grid, TabbedNavigation, withWizard, Button } from '../../components/src';
-import Router from '../../components/src/router'
+import Router from '../../components/proxied-imports/router'
 import { AdUnit, AdUnits, AdSense, HeaderCode, Placements, Services } from './views';
 
 const { HashRouter, Redirect, Route, Switch } = Router
