+++ assets/wizards/engagement/index.js
@@ -18,7 +18,7 @@ import HeaderIcon from '@material-ui/icons/Forum';
  * Internal dependencies.
  */
 import { withWizard } from '../../components/src';
-import Router from '../../components/src/router'
+import Router from '../../components/proxied-imports/router'
 import {
 	Commenting,
 	CommentingDisqus,
